Added a CSS file to store all CSS
You can read the styling.css comment at the top if this didn't make sense.

I figured it would be better to have 1 css file rather than one for each page because it doesn't seem needed and it would be better organized. Better organized as long as we properly comment what each item is for.

You can link it to the specific file you are on with this in the header: <link rel="stylesheet" href="styling.css">