Updated/built out the userProfile page along with other additions
Updated the userProfile page to allow navigation to different pages such as library, bookmarked, etc. Everything is within this file not spread out into others like bookmarked.html. There is another way of doing this that may be better, it just depends on the amount and complexity of data needing to be displayed/used.

Added a profile section within the userProfile that displays their profile picture (larger) and a description. We can add the number rated and other stuff later on. Let me know what you think too. We have a problem with empty space which we will have to discuss and find a solution for.

Changed the placeholder for the search bar on both index and userProfile pages.

I also made it so Logo was clickable both in the userProfile and index files (they operate differently, but once we get a picture in it's an easy switch)