Implemented sign-in/create account
This is from when we were working yesterday.

I built out the sign-in/create account feature. This is accessable either by clicking the sign-in button to the left of the profile icon, or when you try to make a review and are not signed in or don't have an account.

It would be cool let on to make it so once signed-in it changes to log-out

Lastly I say this just incase anything happens, I didn't pull the changes you all made yesturday because I already had changes I had to commit and it said, "unable to pull when changes are in your branch" so hopefully when I commit this your content isn't removed.