I was trying to work on the profile picture
I took out a lot of the implementations because it was causing issues with other update fields in the profile. The main reason for push is so the loadLibrary is added incase I deleted it from main.

An issue i was having with the profile picture was that the image entry was too large. I got the error "PayloadTooLargeError". I tried finding some work around but to no evail. Im going to try again tomorrow before our presentation. worse case I can explain the issue I was having