Ensured the user has to be logged in before making a review
Now if a user goes to make a review they have to be logged in. If they aren't they are re-directed to the log-in.

I also ensured that once they log in they are directed to the profile page.

Lastly I am able to make a review for any piece of media except for books as we knew before. not sure why that is but something we should look at